Update 1066_pares_impares_positivos_negativos.py
I changed float to int in the input.

+++ 1066_pares_impares_positivos_negativos.py
@@ -1,5 +1,4 @@
-#The numbers should be entered with spaces to distinguish them as separate.
-n = list(map(float, input().split()))
+n = list(map(int, input().split()))
 pos = 0
 neg = 0
 par = 0
